+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/page.tsx
@@ -4,7 +4,7 @@ import { Metadata } from "next";
 import { getAppSetting } from "@lib/appSettings";
 import { FormProfile } from "./components/FormProfile";
 import { getFormattedDownloadableTemplateVersions } from "@lib/templates/versioning/queries/getDownloadableTemplateVersions";
-import { DownloadableTemplateVersion } from "@lib/templates/versioning/downloadableTemplateVersion";
+import { type DownloadableTemplateVersion } from "@lib/templates/types";
 
 export async function generateMetadata(props: {
   params: Promise<{ locale: string }>;
